requirements.txt
black~=19.10b0
attrs~=19.3.0
Flask~=1.1.2
pymongo~=3.10.1
pytest~=5.4.3